requirements.txt
requests
Pillow
fastapi
